requirements.txt
python-dotenv==1.2.1
requests==2.32.3
dnspython==2.7.0